+++ NEWS.md
@@ -1,3 +1,8 @@
+# mlr3tuningspaces 0.1.1.9000
+
+* feat: `as_search_space()` function to create search spaces from `TuningSpace` objects. Allows to pass
+  `TuningSpace` objects as `search_space` in `TuningInstanceSingleCrit` and `TuningInstanceMultiCrit`.
+
 # mlr3tuningspaces 0.1.1
 
 * Fix `xgboost.default` search space.